Fix forge weirdness, add publishers, cleanup
Had the most frustrating experience with JIJ. Apparently, using the version catalog MinimalExternalModuleDependency does not work with the extension block added to the outside implementation config. It has to be added on jarJar. Works fine if it's a String though.
Also had fun with mappings, forgot to tell it that it was using namedElements.
Other cleanup, adding configureEach, copyIdeResources, test config extends compileOnly, etc.